Add modules folder and organize the code
